Move MacroProgressView to its own component
